Remove redundant json options
Previously we allow filebeat json options to be customised, but this
lead to significant coupling because the length of the options hash
was used as a signal to enable json processing, even if the options
were unchanged from the defaults. This replaces the options hash with
a simple boolean, since we always set the same options wherever they
were passed in - except for one instance, which appears to be a bug
where the actual intention was to disable the parsing entirely.