Remove hiera of decommissioned Carrenza env
Following on from the decommissioning of the integration and staging
environment of GOV.UK in carrenza, we delete the hiera of these 2
enviroment.

We modify the rake task that checks consistency between Carrenza and AWS
to no longer checks for the removed integration and staging environment.

Future Work: remove Carrenza production and common hiera along with
consistency rake task after Carrenza production is decommissioned.

Refs:
1. [trello card](https://trello.com/c/ITbohL5O/3890-remove-hiera-of-carrenza-integration-and-staging-from-govuk-puppet)
2. [govuk-secrets pr](https://github.com/alphagov/govuk-secrets/pull/1072)