Expose GOVUK_DATA_SYNC_PERIOD ENV variable for govuk_app_config
This is to enable https://github.com/alphagov/govuk_app_config/pull/160,
which ignores certain exceptions if they occur during the data
sync period. We want to set this via Puppet so that it is
configured in only one repo. In the next commit I will DRY up the
values so that they're only set in one place, but for now, there
is precedent for duplicating the values - see
https://github.com/alphagov/govuk-puppet/blob/5b1dc13ea3c7180318da284fe1dbe90a14460234/modules/monitoring/manifests/contacts.pp#L57-L59

We only want to define GOVUK_DATA_SYNC_PERIOD on Integration and
Staging. If we hit an error on Production during the data sync
period, we still want to log it to Sentry.

Trello: https://trello.com/c/rAlvGlSp/2123-5-ignore-data-sync-errors-in-govukappconfig