Move Jenkins Deploy DNS to AWS Production
Currently Deploy_DNS job resides on the Jenkins machine in Carrenza
Production. As Carrenza is going to be decommissioned, we need to move
this job to the Jenkins in AWS production.

The value of `govuk_jenkins::jobs::deploy_dns::gce_project_id` hiera was
set to be the same as in Carrenza Production.

Immediate work to follow:
1. verify that secrets are correct for AWS production
2. test job
3. update documentation and gds-cli
4. retire deploy_dns in Carrenza after some waiting time.

Refs:
1. [Trello Card]
(https://trello.com/c/ti7AGfnE/3894-move-jenkins-deploy-dns-to-aws-production)