Monitor latency on email delivery_* and send_email_* queues
In https://github.com/alphagov/email-alert-api/pull/1438 we are renaming
these queues. This sets up monitoring for the new queues so we can
transfer seamlessly to the new metrics as the app changes.

This should be deployed at a time close to the Email Alert API change as
they will show as pending alerts until Email Alert API starts writing
this data to graphite.