Retry success_passive_check commands until successful
Icinga is reporting some jobs to have the status 'freshness threshold exceeded',
for some jobs which have succeeded. This indicates that the success_passive_check
is not successfully telling Icinga that the job has succeeded.

This retries the success_passive_check until it passes. We could also resend the
message a bunch of times.

If we were not moving away form Icinga soon, we could check out this rewrite of
send_nsca which might be more reliable: https://github.com/Yelp/send_nsca.