Reduce CKAN alert thresholds
In a recent incident CKAN was down but because traffic to the app was
low we didn't get alert until significant numbers of users started using
the app.

The current default critical threshold is 0.1, and during the incident
the error rate reached 0.13 at 10:18, which was 8 hours after CKAN
stopped working.