Migrate CDN Jenkins jobs to AWS Production
As part of the shutdown of Carrenza Production, we need to migrate the
CDN Jenkins jobs to AWS Production.

We use the same list of CDN services for these jobs as in Carrenza
Production.

Future work will include testing the jobs, updating the docs and then
remove the old jobs from Carrenza Production.

Refs:
1. [Trello Card]
(https://trello.com/c/L3325c8c/3900-migrate-cdn-jenkins-jobs-to-aws-production)