Remove check for old AWS IAM keys
Previously we added this check based on a recommendation in AWS
Trusted Advisor. We later decided that this alert has a higher
cost than the value we get from actioning it:

- The number of keys to rotate is high, and the effort required to
rotate a single key is also high.

- Historically (over a year) 2ndline has not had the time to action
the majority of the keys highlighted by this alert.

- We do not believe there is a significant security risk to justify
the effort required in having this alert.

- This alert represents only one of the recommendations in AWS
Trusted Advisor.

In general, rotating keys after 90 days still means there are 90
days where a leaked key may be exploited. We have many issues with
the way we manage secrets, and we should think about resolving them
comphrehensively, rather than with individual alerts.