Focus on sustained 5xx for origin
Previously we would alert for 5xx spikes, which are generally
unactionable (they resolve by themselves). This applies a 5 minute
moving median average to the metric for this alert, which matches
the behaviour of the 5xx alert for individual app proxies.

I've also made the "from" values the same between the in-office and
on-call alerts, as it's unclear why these should be different.