Stop obsessing about established connections
Previously we alerted as soon as the established connections exceeded
the configured threshold, which has lead to lots of flashing alerts.
While it's important to know if an app is receiving more requests than
it can handle, this is only a problem if it's sustained, so we should
apply some averaging to reflect that.