Prevent auto-reboot for monitoring machines
Previously we risked losing alerts if the auto-reboot failed overnight,
so we wouldn't get notified of critical issues with the platform. In
addition, the auto-reboot (at midnight) coincides with the other jobs
like data sync, so we get a load of alerts the next day because the
job was unable to post its status to Icinga. This changes the reboot
settings so that they must be done manually.