Use separate metric for 5xx warnings and criticals
Previously we had to compromise between a knowing about a potential
issue and seeing red "CRITICAL"s when we deem it to actually be a
problem. Averaging over a longer period helps remove false alarms,
but could mask a creeping issue. This separates the concerns by using
a different metric for each purpose.

Note that the "from" settings are used to specify how many data points
to retrieve, which is then used to compute an overall overage.