Limit collection of interface stats
Previously we collected stats for all interfaces, which lead to a
massive increase in data stored for our CI agents, due to their
extensive use of (Docker) veth interfaces. Since we only really
need stats about the physical interfaces, this applies filtering
to the plugin config [1] to restrict the data collected.

[1]: https://collectd.org/documentation/manpages/collectd.conf.5.shtml#plugin_interface