Add authenticated Docker registry for AWS Integration
We enable authenticated puppet docker registry for the AWS Integration
environment. This is to mitigated against the new Docker limits on
docker pulls for anonymous/unpaid accounts.

Future work:
If all goes well, install in other environments.

Ref:
1. [Trello Card]
(https://trello.com/c/M3zyF5kM/3898-identify-govuk-components-and-jenkins-jobs-which-pull-images-from-docker-hub-investigate-an-option-to-enable-authentication-for)
2. [secrets pr]
(https://github.com/alphagov/govuk-secrets/pull/1081)