Only alert for persistent 5xx errors
Previously we applied a moving median average to various 5xx metrics,
which helps to dampen spikes over a short period. Icinga also does its
own (mean) averaging with the "from" param e.g. a "from" value of 8
means we average over three 5 min windows.

However, we can still get false alarms if we have a short period of high
5xx spikes. Increasing the window size is not ideal because this further
masks the data for the purpose of a warning. A compromise is to increase
the "from" value, so that we require more persistence for 5xx errors to
trigger a warning or a critical alert.

This increases the "from" param to 15 minutes for all 5xx alerts, so
that the windowed/de-spiked metric must either:

- Be higher over a shorter period of time; or
- Be over the threshold for a longer period of time