Fix quantities of Sentry errors reported in Grafana
The previous dashboard under-reported the number of Sentry errors
by about 80%, as Graphite uses "events per second for the time
interval, usually 5 seconds. This means that a value of 0.2
represents a single event in a 5 second interval". See
https://docs.publishing.service.gov.uk/manual/graphite-and-deployment-dashboards.html#hitcount

Using `hitcount` converts the time interval data into what we
expect, i.e. an integer value representing the number of events
in the given time period.

I've also taken the opportunity to remove the '30 day average'
line, which isn't hugely useful, and to reduce the time scale from
90 days to 30, which is almost as useful and is more performant.

Now that we've brought the rate-limiting down, I've set stricter
thresholds of rate-limit too, so we need fewer errors to be rate
limited before 'Failure rate' is highlighted orange or red.

The differences can be seen below.

-## Production

-### Before
![prod-before](https://user-images.githubusercontent.com/5111927/98029099-ec6aee80-1e06-11eb-8ec5-4d2680113c7d.png)

-### After
![prod-after](https://user-images.githubusercontent.com/5111927/98029095-ebd25800-1e06-11eb-8c1c-9e24639a4f7e.png)

-## Staging

-### Before
![staging-before](https://user-images.githubusercontent.com/5111927/98029104-ed038500-1e06-11eb-9ae8-74f0f06f0237.png)

-### After
![staging after](https://user-images.githubusercontent.com/5111927/98029621-b24e1c80-1e07-11eb-8d66-a5f5e2b1dca3.png)