Wait even longer to alert about established conns
Previously we applied a moving median average to the metric for this
alert, in order to filter out spikes [1]. However, it's still possible
to get false alarms due to temporary load that lasts a few minutes,
but doesn't represent an issue: the app recovers when as the load dies
down. This extends the data window for this alert from 5 minutes (the
default) to 15 minutes, reduced using a mean average [2].

Mean averaging is preferable because it's more responsive to increases
in the underlying metric. This means we get warnings earlier than if we
instead extended the existing median window to 15 minutes. Using the
combination of averages seems to be the best compromise:

- A 5 minute median average to remove one-off spikes
- A 15 minute mean average of that, to dampen short bursts

[1]: https://github.com/alphagov/govuk-puppet/pull/10804
[2]: https://github.com/pyr/check-graphite#how-it-works