Fix fact check alert not firing when it should
Previously the "1" threshold for this alert failed to trigger due
to a windowing issue where the computed average was actually 0.958.
Since the averaging done by Icinga always means it will always take
some time for this alert to reach a value of "1" [1], this reduces
the threshold so we get alerted faster and more reliably.

Note that there is no need for a "warning" with this alert, since
it's highly unlikely to fix itself.

Note that the previous functions around the metric for this alert
were largely unnecessary:

- keepLastValue is redundant since we also have transformNull
- averageSeries is redundant since Icinga does this anyway [1]

[1]: https://github.com/pyr/check-graphite