Fix fact check email alert again
Previously we modified this alert in an attempt to make sure it
would trigger despite the additional averaging applied by Icinga
[1]. The alert does now trigger, but the averaging still means it
only triggers at a warning level. This is another attempt to get
the alert firing at a CRITICAL level, using "summarize" to get a
moving maximum of the metric that will mean the set of datapoints
to average will exceed the CRITICAL threshold.

Note that the CRITICAL threshold still needs to be lower than one,
as the averaging applied by Icinga can still be slightly less than
this, due to the windowing leaving a "gap on the end".

[1]: https://github.com/alphagov/govuk-puppet/pull/10824