Report Icinga alert states as separate metrics
Previously we reported the state of an alert as the value of the
associated metric, which makes little sense as alert states are
not a "quantity". Using a gauge makes it hard to visualise the
resulting stats, because any aggregation on them will conflate
different states for the alert. This changes the approach for
alert metrics by reporting each state of the alert as its own
metric, so that we can distinguish between them.