Create GDS_SSO_* env vars
This change is done to allow the GDS SSO gem to transition to using env
vars that are prefixed with the dependencies name. This has been done to
make it clear that these env vars are intended for GDS SSO. This change
has been done in advance of changing GDS SSO so that the gem can
seamlessly transition to a new version that uses these env vars.

Once all apps are migrated to use a new version of GDS SSO we can then
delete the old OAUTH_ID env vars.