Don't replay traffic to /transition-check/login/callback
The account-manager redirects the user to this path, with some query
params, after logging in.  If the params are invalid, an error gets
thrown.

Replaying traffic to this endpoint is triggering a lot of spurious
errors, because production query params aren't valid in staging, not
staging query params valid in integration.