Expose arbitrarily hidden healthchecks
Previously we noted that the majority of our apps expose their
healthcheck endpoint [1], with the original concern for not
doing so being vague and not representative of the defaults in
place. Of the three apps that still had their healthcheck route
hidden, none of the commits make any mention of this or suggest
why the endpoint should be treated specially. In addition, two
of the three apps aren't even resolvable externally. This removes
the functionality to hide the healthcheck route and thus exposes
it for three inconsistent apps.

[1]: a91df2aa30d20a2a099b9bf843d1c8c159dfe4d2