Add dashboard to monitor aggregate Icinga alerts
This adds a new Grafana dashboard to help us monitor the number of
Icinga alerts we get over time, both in total and at a per-machine
level. Individual teams can use this dashboard to help monitor alerts
for apps they look after, and 2ndline will use it help monitor the
overall health of GOV.UK and the impact this has on 2ndline.

It's worth noting that, while Icinga does provide some reporting
capabilities, these are oriented around the duration of alerts, as
opposed to individual occurrences. We believe that monitoring the
number of alerts is more useful, due to the sheer volume of alerts
we have per week. Icinga's reports are also difficult to use because
they don't aggregate data across alerts, noting that we often have
duplicate alerts defined for each machine an app runs on.