Reduce integration/staging Email Alert API allowed list
This reduces the number of recipients on this list as a first step
towards this configuration parameter being replaced. Many of these
people are no longer working in GOV.UK and it seems unlikely they
want/need these emails.

I've reduced this list down to just developers on the GOV.UK
Notifications team and Jessica who seems to frequently work with Email
Alert API as part of her Brexit work.

Part of the motivation for this was that I need to duplicate this
configuration to allow it to be renamed and as it is an array I can't
use the normal "%{hiera('')}" technique without converting the array
into a string. Thus a much shorter array is easier to duplicate.