Add GOVUK_NOTIFY_RECIPIENTS env var for Email Alert API
This environment variable is intended to replace the collection of
WHITELIST based environment variables for Email Alert API. This change is
motivated by the previous environment variable approach being complicated
and that using "whitelist" terminology is dated. This is part of a stream
of work that will render a number of Email Alert API environment variables
redundant. This environment variable is inserted to allow transitioning
from one set to another.

This environment variable is intended to accept an "*" to indicate that
all email recipients should receive emails from Notify or a
comma-separated list of email addresses of individuals who'll receive
email from Notify. The absence of the environment variable would
indicate no-one receives Notify emails.

To support the list or "*" approach the puppet module accepts an array
or a string for recipients. It then converts any arrays to a string for
the environment variable.

As the input in the common hieradata is an array, I was unable to use
the "%{hiera()}" approach to avoid needing to repeat a variable - doing
that converted the variable to a string. Thus I thought it was simplest
to just repeat the information.

I did initially intend to name this NOTIFY_RECIPIENTS however we seem to
have a rather broad convention of prefixing GOVUK on any Notify
environment variables.