Block finder-frontend cookies when feature flag is disabled
Rails sets a session cookie on every request, even if there's nothing
to put in the session.  This harms cacheability of pages.  Since the
finder-frontend accounts stuff is hidden behind a feature flag, we
want disabling that feature flag to also disable the cookies.

This has the added benefit of sticking a reference to the feature flag
in this code, so we don't forget to remove it when the transition
checker eventually goes away.