Temporarily re-add the 'marksutton' user
Mark got removed in #10878, a little too quickly: we hadn't yet
applied an `ensure => absent` to their class as per the 'Remove
a user from Puppet' instructions:
https://docs.publishing.service.gov.uk/manual/removing-a-user-from-puppet.html

I can see that they still have a home directory on at least one
backend class machine on Integration, so we need to do a little
housekeeping. We'll temporarily add the user back to Puppet with
an `ensure => absent` instruction. The other properties don't seem
to be required:
https://github.com/alphagov/govuk-puppet/blob/master/modules/govuk_user/manifests/init.pp

When this is merged and Puppet is deployed, we should see the home
directory removed, and we can then revert this commit.

While I'm here, I spotted three other orphaned home directories on
Integration, so am applying the same fix to those users too.

Trello: https://trello.com/c/7ySrimBF/1058-leaver-mark-sutton