Set MEMCACHE_SERVERS in integration and staging
We use a gem called Dalli as a memcached client. This gem takes a
MEMCACHE_SERVERS environment variable, which it will use to work out
which memcached server to talk to.

Currently we're not setting this variable, so we're getting its default
value (127.0.0.1:11211).

To make it easy to override, this makes puppet always set the variable.
By default, we use the same default as Dalli, which should not change
the existing behaviour. In environments where we're using a shared
elasticache memcached instance, we can set it to the DNS name (and port)
of the cluster.

In GOV.UK's current infrastructure, there's a Search path set in
/etc/resolv.conf, so `frontend-memcached` will result in a lookup to
`frontend-memcached.production.govuk-internal.digital`, which will be
CNAMED to `frontend-memcached.blue.production.govuk-internal.digital`.

For now, we're going to experiment with a shared cache in integration
and staging. If it works well, we'll roll it out to production.

See https://github.com/alphagov/govuk-aws/pull/1394 for the
infrastructure side to this.