Increase safe values of nagios memory warnings
Some of our servers can take more than this limit.

In https://github.com/alphagov/govuk-puppet/pull/10895 I bumped the search values, as I didn't want the warning to kick in until we'd used half of the machine memory. Unfortunately, this was higher than this limit, which was invisible until I hit production.

Resolves a puppet run error:

> Error 400 on SERVER: validate_integer(): Expected 16000 to be smaller or equal to 12000, got 16000. at /usr/share/puppet/production/current/modules/govuk/manifests/app/config.pp:143

I _think_ this is an arbitrary limit, to try to ensure that we don't have too many zeroes.