Add frontend memcache servers to production
We've completed a round of testing on the memcache configuration and
concluded that we think it's beneficial to apply this to production.
I've done this by only defining this for the frontend node and not
setting anything in the draft stack (where it fall back to default).
There was some confusion in the reviews regarding the localhost IP
address so I've removed that. As I understand this was done to make it
easy to switch the service off in an environment, however I believe we
can achieve the same by setting a hieradata of null, for example:
`govuk::apps::collections::memcache_servers: ~`.

In testing we confirmed that this shared memcache instance dropped
response times by around 50% compared to having to request a postcode
from mapit.

We also found that when hit with a load that was enough to overwhelm our
app instances we only seemed to utilise about 1% of the CPU and
available memory was still very high. Finally we tested how Rails copes
with a failure and this seemed to be graceful, with an application just
executing the non-cached code if the resource is unavailable.