Specify 1 attempt for PagerDuty drill
The PagerDuty drill works by triggering the creation of a file at 10:00
on a Wednesday then removing that file at 10:15. An Icinga check becomes
critical when this file exists.

PagerDuty is alerted when this check becomes "hard". We are currently
falling back to a default of 4 critical attempts before the alert is
considered "hard", so occasionally PagerDuty will not be alerted if the
4 checks don't finish in the 15 minute period.

We should change this to 1 attempt before becoming "hard", so PagerDuty
will always get alerted when this file is found after 10:00.