Use normal default vhost for Content Store
This switches the default vhost config config so it's the same
as that used by other apps, which is still compatible with the
current LB config, as the following diff on Staging shows:

 server {
   listen 80 default_server;

-  rewrite ^/_healthcheck$ /healthcheck;
-
   # Send the Strict-Transport-Security header
   include /etc/nginx/add-sts.conf;
   # Required for Layer 7 Application Load Balancer
-  location = /healthcheck {
-    proxy_set_header Host content-store;
-    proxy_pass http://localhost:3068;
+  location = /_healthcheck {
+    return 200;
   }
+  include /etc/nginx/lb_healthchecks.conf;

   location / {
     return 500 '{"error": "Fell through to default vhost"}\n';
   }
+
 }
+

Note that the "/healthcheck" endpoint is redundant, as the
"/_.." one is used by the LB [1]. While the new version just
checks that Nginx is up, we want to follow up this change with
one to the LB config, so that it uses the following endpoint in
the new "/etc/nginx/lb_healthchecks.conf" file.

location /_healthcheck_content-store {
  proxy_pass http://content-store-proxy/healthcheck;
}

The net result will be a more granular healthcheck for Content
Store, as opposed to the current situation where the LB simply
checks whether the app is responding.

[1]: https://github.com/alphagov/govuk-aws/pull/1402/files#diff-069b7aaa2455edca6b507407de527eba4e4374d04b01584889519b6c5d6a4290R957
[2]: https://github.com/alphagov/govuk-aws/pull/1396