Add prometheus data source to Grafana
It looks like this can work the same way as graphite does - with
https://prometheus as the URL. This works because /etc/resolve.conf sets
a search path of integration.govuk-internal.digital.
prometheus.integration.govuk-internal.digital is a CNAME to
prometheus.blue.integration.govuk-internal.digital, which points to
Prometheus' internal load balancer.

I'm not sure how Grafana is happy making an HTTPS connection to this
URL, becauce the certificate is definitely not valid for
https://prometheus. But I've tested it and it seems happy (just as it is
with graphite), and not having to pass the URL in as a variable makes
our lives significantly simpler.