Ensure Sentry Puppet resources are removed prior to deletion
In #10919, we're removing Sentry error-reporting for failed Puppet builds,
but all this does is tell Puppet not to manage those resources
anymore. It doesn't actually remove them.

We need to tell Puppet to remove the resources before we can
safely remove the resource configurations.

See https://serverfault.com/questions/455657/how-to-revert-a-configuration-with-puppet

Trello: https://trello.com/c/1pAh0QpK/2294-3-remove-govuk-puppet-project-from-sentry

Co-authored-by: Thomas Leese <thomas.leese@digital.cabinet-office.gov.uk>